refactor: move Eater as interface of livetraffic
